Refactor dev module structure and update imports
Moved and renamed several files in the dev module for improved organization, including controllers and transformers. Deleted unused core/conversion.py and core/contollers/__init__.py. Updated import paths in audioeditor.py and server/api.py to reflect new structure. Added initial content to test.ipynb and minor README update.
