--- src/bioamla/dev/core/contollers/__init__.py
@@ -1,18 +1,0 @@
-"""
-Audio Processing Controllers Module
-
-This package contains controller classes that coordinate and manage complex
-audio processing workflows. Controllers act as orchestrators that combine
-multiple audio processing operations and manage the interaction between
-different components of the audio editing system.
-
-The controllers module provides high-level interfaces for:
-- Coordinating multi-step audio processing pipelines
-- Managing workflow state and progress tracking  
-- Handling complex audio editing operations
-- Orchestrating interactions between audio processing components
-
-This module serves as the control layer that sits between the user interface
-and the core audio processing functionality, providing a clean separation
-of concerns and enabling complex audio editing workflows.
-"""
